+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### New features and improvements
 
 - App: Added a button to copy the version string from the About window.
+- App: Added option to automatically load most recently loaded game disc image on startup.
 - App: Auto-center About window whenever it is opened.
 - Build: Nightly builds are now available [here](https://github.com/StrikerX3/Ymir/releases/latest-nightly).
 - Core: Improve manual reset event performance by using OS-specific implementations based on cppcoro.